Refine native inspector and retire standalone research records
Unify inline metadata editing, attachment Quick Look, note notifications, link context presentation, and transient line navigation feedback. Remove dedicated Records UI, storage, search, and MCP contracts in favor of ordinary Works notes.

Update specifications, architecture, skills, fixtures, and owning tests; normalize native typography and color ownership and repair stale interface assertions.

+++ Tests/ScholiumApplicationTests/ScholiumMCPServerTests.swift
@@ -10,7 +10,7 @@ struct ScholiumMCPServerTests {
         let recorder = MCPRequestRecorder()
         let server = ScholiumMCPServer { request in
             await recorder.record(request)
-            return .object(["schema_version": .integer(2), "status": .string("ok")])
+            return .object(["schema_version": .integer(3), "status": .string("ok")])
         }
 
         let initialized = try await rpc(
@@ -29,7 +29,7 @@ struct ScholiumMCPServerTests {
         let tools = try #require(listResult["tools"] as? [[String: Any]])
         #expect(tools.compactMap { $0["name"] as? String } ==
             ScholiumMCPToolName.allCases.map(\.rawValue))
-        #expect(tools.count == 10)
+        #expect(tools.count == 7)
         for tool in tools {
             let schema = try object(tool["inputSchema"])
             #expect(schema["additionalProperties"] as? Bool == false)
@@ -53,7 +53,7 @@ struct ScholiumMCPServerTests {
         let server = ScholiumMCPServer { request in
             await recorder.record(request)
             return .object([
-                "schema_version": .integer(2),
+                "schema_version": .integer(3),
                 "status": .string("ok"),
                 "current": .bool(false),
             ])
@@ -98,7 +98,7 @@ struct ScholiumMCPServerTests {
         let result = try object(response["result"])
         #expect(result["isError"] as? Bool == true)
         let structured = try object(result["structuredContent"])
-        #expect(structured["schema_version"] as? Int == 2)
+        #expect(structured["schema_version"] as? Int == 3)
         #expect(structured["status"] as? String == "failed")
         #expect(structured["code"] as? String == "workspace_not_ready")
         #expect(structured["recovery"] as? String == "Open one Triptych.")
